+++ cloudsmith_cli/cli/commands/__init__.py
@@ -5,6 +5,7 @@ from __future__ import absolute_import, print_function, unicode_literals
 from . import check  # noqa
 from . import copy  # noqa
 from . import delete  # noqa
+from . import dependencies  # noqa
 from . import docs  # noqa
 from . import entitlements  # noqa
 from . import help_  # noqa