+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added the `cloudsmith dependencies` sub-command, to list package dependencies.
+
 ## [0.30.2] - 2021-12-21
 
 ### Fixed